+++ verik-compiler/src/main/kotlin/io/verik/compiler/main/StageSequencer.kt
@@ -33,6 +33,7 @@ import io.verik.compiler.check.post.PostNameCheckerStage
 import io.verik.compiler.check.post.StatementCheckerStage
 import io.verik.compiler.check.post.UntransformedElementCheckerStage
 import io.verik.compiler.check.post.UntransformedReferenceCheckerStage
+import io.verik.compiler.check.pre.ConfigCheckerStage
 import io.verik.compiler.check.pre.FileAnnotationCheckerStage
 import io.verik.compiler.check.pre.ImportDirectiveCheckerStage
 import io.verik.compiler.check.pre.PreNameCheckerStage
@@ -108,6 +109,7 @@ object StageSequencer {
         stageSequence.add(StageType.PARSE, KotlinEnvironmentBuilderStage)
         stageSequence.add(StageType.PARSE, KotlinCompilerParserStage)
 
+        stageSequence.add(StageType.PRE_CHECK, ConfigCheckerStage)
         stageSequence.add(StageType.PRE_CHECK, FileAnnotationCheckerStage)
         stageSequence.add(StageType.PRE_CHECK, UnsupportedElementCheckerStage)
         stageSequence.add(StageType.PRE_CHECK, UnsupportedModifierCheckerStage)