+++ verik-plugin/src/main/kotlin/io/verik/plugin/VerikPluginExtension.kt
@@ -19,13 +19,14 @@ package io.verik.plugin
 abstract class VerikPluginExtension {
 
     var timescale: String = "1ns / 1ns"
+    var entryPoints: ArrayList<String> = ArrayList()
     var enableDeadCodeElimination: Boolean = true
     var labelSourceLocations: Boolean = true
     var enableLineDirective: Boolean = true
     var indentLength: Int = 4
     var wrapLength: Int = 120
-    var suppressedWarnings: List<String> = listOf()
-    var promotedWarnings: List<String> = listOf()
+    var suppressedWarnings: ArrayList<String> = ArrayList()
+    var promotedWarnings: ArrayList<String> = ArrayList()
     var maxErrorCount: Int = 20
     var debug: Boolean = false
 }