+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - `Verik` annotation to enable inspections by IntelliJ plugin.
 - File annotation checker issues warnings if `Verik` annotation is missing.
 - Consolidate interpreter checks to midCheck.
+- Configuration validity checks.
+
+### Changed
+- Specify entry points in `VerikPluginExtension`.
 
 ## [0.1.10]
 ### Added